+++ plugins/remark.ts
@@ -16,7 +16,10 @@ export interface Options {
   /** List of extensions this plugin applies to */
   extensions: string[];
 
-  /** List of remark plugins to use */
+  /**
+   * List of remark plugins to use
+   * @default `[remarkGfm]`
+   */
   remarkPlugins?: unknown[];
 
   /** List of rehype plugins to use */